Rename ScrollProvider to useScrollListener and move to hooks folder
